requirements.txt
tensorflow==2.6.0
numpy==1.21.2
opencv-python==4.5.3.56
keras==2.6.0
Pillow==8.2.0
